+++ main.py
@@ -6,9 +6,9 @@ from pathlib import Path
 from pprint import pprint
 
 from app.datamodels.models import JobInfo
-from app.scoring import score_resume, summarize_gaps, suggest_edits
-from app.scoring.prompt_extraction import check_and_extract, extract_tailoring
-from app.scoring.success_prediction import calculate_interview_chance
+from app.workflow import score_resume, summarize_gaps, suggest_edits
+from app.workflow.prompt_extraction import check_and_extract, extract_tailoring
+from app.workflow.success_prediction import calculate_interview_chance
 
 logging.basicConfig(
     level=logging.INFO,
